+++ assets/javascripts/kitten/components/cards/vertical-card-with-action/vertical-card-with-action.stories.js
@@ -13,7 +13,8 @@ const buttonModifierOptions = {
   Oxygen: 'oxygen',
   Checked: 'checked',
   Copper: 'copper',
-  Boron : 'boron',
+  Boron: 'boron',
+  Neon: 'neon', 
 }
 
 export const Default = () => (