+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `VerticalProgress`: Fix `z-index` value.
+
 ## [3.13.1] - 2021-05-20
  
 Fixes: