+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: `VerticalProgress`: Fix `z-index` value.
+- Feature: Add `neon` modifier to Button component.
 
 ## [3.13.1] - 2021-05-20